+++ apps/web/src/lib/components/HomePage.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+	import Navigation from './Navigation.svelte';
+</script>
+
+<Navigation />
+<div>
+  Home Page
+</div>
